Introduce link tracking
My use of wireguard largely incorporates site-to-site tunnels with
routing information exchanged over misc protocols by misc routing
suites. This is a match made in heaven, with one exception: currently
there is no simple way for a routing daemon to determine whether or
not a wireguard peer is actually up as this is not exposed to the
netdev layer at all, leaving down detection to the hello/dead timers
of routing protocols employed (or, alternatively, bfd). Futhermore
the local link used between wg peers always stays up, leaving the
kernel to believe it always has a link-local route to all the
addresses in the subnet or subnets that the wg device is configured
with. This quickly gets painful and all of it would go away instantly
if only wireguard was able to signal the loss of peer connectivity
by simply marking the wg interface as down (ie simulate loss of
link).

This commit attempts to solve this, by introducing a 'link tracking'
concept to wireguard. It works by hooking into the socket/handshake
methods of wireguard to determine whether a wireguard peer is 'up'
(meaning there is a valid handshake present) and proceeds to force
down the lower/carrier layer of the wg interface if the peer is
determined to be down.

The feature has been dubbed 'linktrack' but I'm sure there is a
better name, suggestions welcome.

It comes with two modes; "any" wherein the wg interface is considered
up as long at least one peer is active, and "all" wherein the
interface is considered down unless all peers are active.

To configure link tracking using wg:
  `wg set wg0 linktrack [any|all|off]`

This patch should be considered proof-of-concept quality as I have
not yet evaluated the full potential impact of the changes.